Separate feature initialization into Feature_Loader class
- Create Feature_Loader class to handle orchestration
- Move initialization logic from Feature_Registry to Feature_Loader
- Feature_Registry now only handles storage/retrieval
- Update Plugin class to use Feature_Loader
- Update tests to use Feature_Loader for initialization
