+++ tests/Integration/Features/Example_Feature/Example_FeatureTest.php
@@ -8,6 +8,7 @@
 namespace WordPress\AI\Tests\Integration\Features\Example_Feature;
 
 use WordPress\AI\Feature_Registry;
+use WordPress\AI\Feature_Loader;
 use WordPress\AI\Features\Example_Feature\Example_Feature;
 use WP_UnitTestCase;
 
@@ -26,7 +27,9 @@ class Example_FeatureTest extends WP_UnitTestCase {
 		parent::setUp();
 
 		$registry = wp_ai_feature_registry();
-		$registry->initialize_features();
+		$loader   = new Feature_Loader( $registry );
+		$loader->register_default_features();
+		$loader->initialize_features();
 
 		$feature = $registry->get_feature( 'example-feature' );
 		$this->assertInstanceOf( Example_Feature::class, $feature, 'Example feature should be registered in the registry.' );
